requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
flake8==3.5.0
flake8-polyfill==1.0.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.0
more-itertools==4.2.0
pep8-naming==0.7.0
pluggy==0.6.0
py==1.5.3
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
pymongo==3.6.1
pytest==3.6.0
python-dotenv==0.8.2
six==1.11.0
Werkzeug==0.14.1
WTForms==2.1
